Storing `page_ref` inside `kmem` instead of `run`
`struct run` is a linked list that stores the free pages that we can use
to allocate new pages. I don't think we should put the page reference
counter there. Instead, use the `kmem`, where kernel memory is.
In order to accomplish this, I create a table based on the number of
frames we have available: `pg_refs[PHYS_STOP >> PGSHIFT]`, this should
divide the ammount of physical mem avaiable by the page size, which
should calculate the number of physical pages/frames we can use.